requirements.txt
lief
pyaxmlparser
pycryptodome
